[TASK] Reintroduce missing core field configured by static_info_tables
Current versions of both l10nmgr and static_info_tables rely on that
field, which has been removed from the core recently.

Although this field is still configured in static_info_tables TCA, the
extension does not make sure it really exists, so l10nmgr will at least
deal with that issue until it is solved.

Static info tables should be made obsolete with the "make it right"
release of l10nmgr 10 though.