[FEATURE] Add responsible customer name as optional header field
Feature needs to be enabled by extension configuration to avoid
unwanted information disclosure.

Default for CLI based exports will be the real name of the CLI user.
Default for CLI based exports can be overriden by a CLI parameter "customer"
Customer name for backend based exports will be the real name of the BE user