[BUGFIX] Prevent export of not needed file metadata
During export the script iterates through all records on the page and
decides wether to select them for the actual export.
However for sys_file_reference records files will always be selected
no matter if the reference record should be exported or not.

This can lead to loads of not needed sys_file_metadata records in the
export.

Bonus: The if-else-construct when selecting records to export has identical
bodies in the if and the else-if part. So the structure can be simplified by
creating one better if statement.