[BUGFIX] Fix self closing br tags
The TYPO3 CKEditor saves br-tags as selfclosing tags with a space 
inbetween: `<p>Some text<br /> with break</p>`
During the import phase the space gets removed by the 
XmlTool::xmlRecompileFromStructValArray() method. 
 As a result the html is saved 
without the space in the TYPO3 db: `<p>Some text<br/> with break</p>`
When an editor then tries to edit the text,
 the br gets escaped: `<p>Some text&lt;br/&gt; with break</p>` 
and will be shown in the frontend after save.